+++ src/main/java/org/osiam/addons/administration/model/command/EditGroupCommand.java
@@ -1,48 +1,32 @@
 package org.osiam.addons.administration.model.command;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import org.hibernate.validator.constraints.NotEmpty;
 import org.osiam.resources.scim.Group;
-import org.osiam.resources.scim.MemberRef;
 import org.osiam.resources.scim.UpdateGroup;
 
 /**
  * Command object for the group update view.
  */
-public class UpdateGroupCommand {
+public class EditGroupCommand {
     private String id;
 
     @NotEmpty
     private String displayName;
     private String externalId;
 
-    private List<String> memberIds = new ArrayList<String>();
-
     /**
      * Creates a new UpdateGroupCommand based on the given {@link Group}.
      *
-     * @param group
-     *            the user
+     * @param group the group
      */
-    public UpdateGroupCommand(Group group) {
+    public EditGroupCommand(Group group) {
         setId(group.getId());
 
         setExternalId(group.getExternalId());
         setDisplayName(group.getDisplayName());
-
-        if (group.getMembers() != null) {
-            for (MemberRef ref : group.getMembers()) {
-                memberIds.add(ref.getValue());
-            }
-        }
     }
 
-    /**
-     * Creates a new UpdateGroupCommand.
-     */
-    public UpdateGroupCommand() {
+    EditGroupCommand() {
     }
 
     /**
@@ -57,19 +41,10 @@ public class UpdateGroupCommand {
     /**
      * Sets the displayname.
      *
-     * @param displayName
-     *            the displayname to set
+     * @param displayName the displayname to set
      */
     public void setDisplayName(String displayName) {
         this.displayName = displayName;
-    }
-
-    public List<String> getMemberIds() {
-        return memberIds;
-    }
-
-    public void setMemberIds(List<String> memberIds) {
-        this.memberIds = memberIds;
     }
 
     public String getId() {
@@ -102,12 +77,6 @@ public class UpdateGroupCommand {
         } else {
             builder.updateExternalId(getExternalId());
         }
-
-        builder.deleteMembers();
-        for (String member : memberIds) {
-            builder.addMember(member);
-        }
-
         return builder.build();
     }
 }